Use fixed size staging buffer - Refactor classes - Add missing resources' clean up - Track allocated image memory - Update 0.5.0-dev+14 - Sync 3ac10f3

+++ src/main/java/net/vulkanmod/render/chunk/buffer/UploadManager.java
@@ -85,7 +85,7 @@ public class UploadManager {
         VkCommandBuffer commandBuffer = this.commandBuffer.getHandle();
 
         StagingBuffer stagingBuffer = Vulkan.getStagingBuffer();
-        stagingBuffer.copyBuffer((int) bufferSize, src);
+        stagingBuffer.copyBuffer(bufferSize, src);
 
         if (!this.dstBuffers.add(buffer.getId())) {
             try (MemoryStack stack = MemoryStack.stackPush()) {
@@ -112,7 +112,7 @@ public class UploadManager {
         copyBuffer(src, 0, dst, 0, src.getBufferSize());
     }
 
-    public void copyBuffer(Buffer src, int srcOffset, Buffer dst, int dstOffset, int size) {
+    public void copyBuffer(Buffer src, long srcOffset, Buffer dst, long dstOffset, long size) {
         beginCommands();
 
         VkCommandBuffer commandBuffer = this.commandBuffer.getHandle();
