+++ src/main/java/net/vulkanmod/vulkan/memory/VertexBuffer.java
@@ -6,19 +6,18 @@ import static org.lwjgl.vulkan.VK10.VK_BUFFER_USAGE_VERTEX_BUFFER_BIT;
 
 public class VertexBuffer extends Buffer {
 
-    public VertexBuffer(int size) {
+    public VertexBuffer(long size) {
         this(size, MemoryTypes.HOST_MEM);
     }
 
-    public VertexBuffer(int size, MemoryType type) {
+    public VertexBuffer(long size, MemoryType type) {
         super(VK_BUFFER_USAGE_VERTEX_BUFFER_BIT, type);
         this.createBuffer(size);
 
     }
 
     public void copyToVertexBuffer(long vertexSize, long vertexCount, ByteBuffer byteBuffer) {
-        int bufferSize = (int) (vertexSize * vertexCount);
-//        long bufferSize = byteBuffer.limit();
+        long bufferSize = vertexSize * vertexCount;
 
         if(bufferSize > this.bufferSize - this.usedBytes) {
             resizeBuffer((this.bufferSize + bufferSize) * 2);
@@ -30,11 +29,9 @@ public class VertexBuffer extends Buffer {
 
     }
 
-    private void resizeBuffer(int newSize) {
+    private void resizeBuffer(long newSize) {
         MemoryManager.getInstance().addToFreeable(this);
         this.createBuffer(newSize);
-
-//        System.out.println("resized vertexBuffer to: " + newSize);
     }
 
 }
